Remove the use of MAIL_REPLY_TO
This was to set the Reply-To: mail header in messages sent from the
application. This header is what would be used when replying to the
message.

We don't actually want to set this header at all, people should not be
replying to these emails. They come from noreply@dnsandmx.com and any
replies should be sent there which in turn just goes to /dev/null.

So remove all uses of MAIL_REPLY_TO

Signed-off-by: Andrew Clayton <andrew@dnsandmx.com>