config: Set the envelope sender in the example config
Emails from the application come from noreply@dnsandmx.com. Replies to
these emails just end up in /dev/null. What we do want is any error
messages to go somewhere they'll be seen. So we set the envelope sender
(-f) on the sendmail command to a real address.

Show this in the example config file.

Signed-off-by: Andrew Clayton <andrew@dnsandmx.com>