dax/sign_up/: Check for a valid email address
On the main dnsandmx.com page there is an input box to enter an email
address in order to sign up for an account. While this input box is set
with a type of "email" it seems some browsers don't enforce this and
will allow anything to be entered there (yes I have seen this happen).

So add a check in src/url_handlers.c::sign_up to check for a valid email
address and inform the user if what they have entered is invalid.

Signed-off-by: Andrew Clayton <andrew@dnsandmx.com>