templates: mail_fwd_record: Allow a special destination of devnull
It is often useful to have a valid email address that doesn't actually
want to receive email. Such as noreply@...

In order to facilitate this, you can now create an email alias that
points to the destination 'devnull'.

This needs to be devnull and not /dev/null as you can't put the file
target in the virtual alias maps, so we have to create a devnull entry
in /etc/aliases and then devnull will be expanded out to
devnull@<host>.domain and looked up in /etc/aliases where it is pointed
to /dev/null

Signed-off-by: Andrew Clayton <andrew@dnsandmx.com>