templates: sign_up: g_strstrip() the email address
There was an instance where someone had created an account but couldn't
log in. Turns out that they had managed to enter some whitespace at the
beginning of their email address. With browsers that support the 'email'
input type, the browser will catch this.

However, to cater for browsers that don't (and just for good practice
anyway) we should strip off any leading and trailing whitespace from the
email address.

Signed-off-by: Andrew Clayton <andrew@dnsandmx.com>