Add the ability to extract ones domain data as a CSV file
This allows users to download their domain data as a TAB delimited CSV
file. It has the following format

    name\ttype\tcontent\tttl\tprio\tmodified\n

The modified field is taken from the change_date field of the pdns
records table. It is a UNIX timestamp, this won't be populated,for
slave domains.

This also works for disabled domains giving people a viable migration
strategy if they don't wish to continue with our service.

Signed-off-by: Andrew Clayton <andrew@dnsandmx.com>