Add the ability to extract ones domain data as a BIND zone file
This adds a 'Tools' link to the Overview page. From there a user can
select a domain and a format (currently only BIND zone files are there)
and they will get a download of their domain data in BIND zone file
format.

This also works for disabled domains giving people a viable migration
strategy if they don't wish to continue with our service.

Signed-off-by: Andrew Clayton <andrew@dnsandmx.com>