utils: is_domain_editable: Simplify the check
When we check if a domain is editable by a user, we check that the
domain belongs to the user and that it hasn't expired. Previously we
would pull the expired value from the database for that users domain and
check if it was a 0 (not expired).

We can simplify this and do the whole check in the SQL by checking the
expired value is 0 as well that the uid and domain_id match.

If we get a row returned the domain is editable else it's not.

Signed-off-by: Andrew Clayton <andrew@dnsandmx.com>