utils: Re-work is_valid_email_address()
The main issue was that it would not allow an email address if it did
not contain a '.'. However it seems that could be valid e.g

    me@com

We now make sure the address is at least 3 characters long, doesn't end
with a '.' and contains an '@', but doesn't have '@.'

e.g

    @a
    @.com
    .@.
    user@example.com.

would be treated as invalid.

Signed-off-by: Andrew Clayton <andrew@dnsandmx.com>