utils: is_valid_hostname: Use NI_MAXHOST for length check
This function may be checking either a hostname or FQDN, previously we
were using 63 for the max hostname length, but for a FQDN this may be
too short.

Make the length check >= NI_MAXHOST (it includes space for the nul
byte). NI_MAXHOST is defined in netdb.h and is used by functions like
getnameinfo(3).

Signed-off-by: Andrew Clayton <andrew@dnsandmx.com>