url_handlers: Remove possible SQL injection in delete_mail_fwd_record
In /mail_forwarding/ we list the forwarding rules for the given domain,
we also provide a delete button for each record that does a POST to
/delete_mail_fwd_record/

Previously this was passing the domain id and the source forwarding rule
(essentially an email address). This source parameter was just being
used in the SQL DELETE query as is, it would be quite simple for someone
to craft a source= that contained some nasty SQL.

Remove this problem altogether by using the record_id, which is a simple
integer instead of source.

Signed-off-by: Andrew Clayton <andrew@dnsandmx.com>