url_handlers: delete_mail_domain: Fix mail domain deletion
Commit 8abb360 ("url_handlers: Convert delete_mail_domain to
is_users_domain") introduced a subtle bug when deleting mail domains.

We were assigning a pointer to the mail domain type, however the
variable that the pointer was pointing to was being free'd before we'd
finished using the pointer and thus got messed up.

strdupa(3) the 'type' variable to solve this.

Signed-off-by: Andrew Clayton <andrew@dnsandmx.com>