utils: Don't always set the user banner in send_template()
In send_template() we add the users name and last login details to the
page we send out. However there is no point in doing this for the pages
where there is no user session, e.g the /login/ page or the /logout/
page.

This saves on a database query to get the last login details.

Signed-off-by: Andrew Clayton <andrew@dnsandmx.com>