utmp: Log the remote port for the login
This logs the port on the remote end that the connection originated
from. In the case of NAT this will be a port on the NAT device.

The main point of this is to allow some chance of linking a login back
to a specific host behind NAT where the logged port will be a port on
the NAT device which you can use to link back to a host behind the NAT
e.g by looking in /proc/net/ip_conntrack on Linux.

Signed-off-by: Andrew Clayton <andrew@dnsandmx.com>