Makefile: Compile with -fcommon
This was the default up until GCC 10.x, which switched to -fno-common as
default.

Due to some poor decisions on my part this now results in a bunch of
linker errors regarding multiple definitions of variables.

I will fix that up but in the meantime just build with -fcommon so this
will build with recent GCC.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>