Fix problems with the global MYSQL conn variable
There is a global variable defined called 'conn' which holds a pointer
to a MySQL connection. This is used in many places and was perfectly
fine as there was one of these per process and no threads...

... until there was threads.

Commit b30bad0 ("dax: Fix potentially slow logins") introduced the use
of threads to do reverse name lookups. These threads want to use the
global MySQL connection (via make_mysql_safe_string()). That will
invariably lead to _problems_.

I think the simplest thing to do here is to declare the global
connection handler as thread local storage. The threads can then just
set that.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>